fix: refuse protected system subtrees
